Increment payment authorization, Four keys & OAuth and minor improvements
This commit adds support for increment payment authorization related
functionality available for Four accounts. In order to accept API Keys / OAuth
tokens, a new configuration constructor was created for this purpose. The
existing constructor was marked as deprecated in favor of adoption of
the new one. The new implementation does not rely on environment lookup
using the key provided.

Unit tests and integration tests were added to test added functionality
and documentation updated to match the new implementation standard.

Also, this commit fixes minor issues found.

Note: the existing codebase is not fully compatible with Four data model
and specific operations, which means that certain functionalities might
not work as expected when using the SDK using a Four/OAuth
configuration.