Release 1.0.0
This commit represents an effort made by the Checkout team to build a brand new SDK for GO. It ships a new instantiation layer that makes easier a less confusing to set up the SDK. From the point of view of usability, the new version follows similar principles, although major clients are now split/duplicated to reference different data schemes which makes it fully compatible with `default` (NAS) and `Previous` (ABC) account structures. This separation is statically forced during instantiation.

The SDK supports the following modules:

* Alternative Payments Clients & Sources (default/previous)
* Customers (default/previous)
* Disputes (default/previous)
* Instruments (default/previous)
* Payments (default/previous)
* Payments Hosted & Links (default/previous)
* Tokens (default/previous)
* Accounts (default)
* Balances (default)
* Card Metadata (default)
* Forex (default)
* Sessions (default)
* Transfers (default)
* Workflows (default)
* Events (previous)
* Sources (previous)

`README.md` was updated to match the new setup. Code of conduct and licence were reviewed.