Allow Amount to be zero in sessions.PaymentSessionsRequest
This change allows the `Amount` field to be zero so we can use `PaymentSessionsRequest` objects for verifications in `Flow`.

Previously if you had tried to create a payment session request object with an Amount of zero, indicating you're doing a verification, the amount field would be excluded from the request payload going over the wire.  This resulted in the CreatePaymentSession API call in kicking back a `amount_required` error since the Checkout API didn't receive the `amount` field.